Make region link consistent with paas-admin
The previous commit made the region name in the top-right of UAA pages
link to a way to switch region. However the equivalent change to
`paas-admin` [1] had a bug. That bug didn't affect this repo but led
to some style improvements. This commit brings in those style
improvements.

Tested out manually in a browser.

[1] https://github.com/alphagov/paas-admin/pull/269