Add javascript for copying button to clipboard
We only want the button to appear when javascript is enabled, we could
either do this with css, or with javascript. Given the rest of the
interaction is with JS then JS seems appropriate.

This adds a secondary (grey) button which appears below the code, if the
user has javascript enabled. The button is focused and, when clicked,
uses the navigator api to write the code to the user's clipboard.

Most modern browsers support the navigator API, and if the user's
browser does not then the button will not appear.

The <p> tag is added to ensure the code is rendered as a block, and the
button appears below.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>