Fix the SVG render in Microsoft Edge
We're passing in the Crown logo on our pages with an inline SVG block.
This is great, as less resources are called in to render the page the
way we want it to be. SVGs have the fallback URL, just in case a browser
is not capable of supporting the embedded code.

This causes us two issues:

1. Microsoft Edge was attempting to load both resources (SVG and the
fallback PNG)
2. The PNG was returning a 302, which redirects the user to the login
page

The Design System team has addressed the issue number 1 and Ollie did an
amazing job of explaining the problem in the [commit message][1].

This is more of an accessibility and reliability issue.

The one we've been facing is number 2.

We're generating a CSRF token, on each load of the forms element for
security purposes. Each time the page is loaded, it's passing the token
into the form as well as setting up a cookie for the user, just to
validate that token elsewhere later.

If the page is being loaded twice, that means:
- a token is being passed in HTML and set as a cookie
- a token is being set as a cookie - reset, no longer matching HTML

Therefore after a form submission, the code that was sent via the form,
does not match the code that has been sent via the cookie.

This is weirdly solved, by providing a fallback URL which will in fact
resolve to a PNG or a 404, instead of more HTML.

Reconstruction, by Richard Towers:

* The user visits a page served by UAA (say `/login`, or
  `/accept/invitation` - anything that returns a form)
* The page contains an SVG, with an image fallback
* Most browsers are happy with the SVG and don't request the image, but
  IE / Edge quirkily request both
* The path to the image is wrong, but because the user is not logged in
  it returns a 302 to `/login` instead of a 404
* The browser follows the 302 redirect, now looking for the image at
  `/login`
* Because the response for `/login` is a form, it has a `Set-Cookie`
  header that overwrites the csrf token for the form the user is
  actually looking at.
* When the user submits the form, the token in the HTML doesn't match
  the token in their cookie, so their request is denied.

[1]: https://github.com/alphagov/govuk-frontend/pull/1337/commits/84485f589523be4101fd1912b1e2e91af7f9a3bf